GUI refactor, computer game works
