+++ src/edu/ucsb/cs56/projects/games/battleship/view/BattleshipGUI.java
@@ -24,36 +24,20 @@ public class BattleshipGUI extends JFrame{
     private String difficulty = null;
    	private int gameType = 0;
     private int replayType = 0;
-	private int lastMove;
-	private int xLoc;
-	private int yLoc;
-	private int boatToPlace;
+
 	private int boatSpawn;
-    private Color shipColor = Color.DARK_GRAY;
-	private boolean playersTurn = false;
 	private boolean ipEntered = false;
 	private boolean replay = true;
 	private boolean prompt = true;
-	private boolean placeBoats = false;
-	private boolean boatPlaced = false;
-	private boolean horzOrVert = true; //true for horizontal false for verticle
-    private boolean audio = true;
 
     //Public so that other classes can play sound files
-    public URL placeURL;
+
     public URL shotURL;
     public URL missURL;
     public URL winURL;
     public URL loseURL;
-    public URL cantPlaceURL;
-    public Clip clip;
 	
-	//GUI's knowledge bank. Used for GameGrid cell coloring
-	private ArrayList<Integer> playerBoats = new ArrayList<Integer>();
-	private ArrayList<Integer> enemyBoats = new ArrayList<Integer>();
-	private ArrayList<Integer> shots = new ArrayList<Integer>();
-    private ArrayList<ArrayList<Integer>> playerBoatGroups = new ArrayList<ArrayList<Integer>>();
-	public ArrayList<Integer> getEnemyBoats() {return this.enemyBoats;}
+    private JPanel audioPanel = new JPanel();
 
 	//GUI Texts
    	private JLabel title = new JLabel("Battleship",JLabel.CENTER);
@@ -108,9 +92,7 @@ public class BattleshipGUI extends JFrame{
     //Game board component
     private GameGrid board = new GameGrid();
 
-    //Audio muted/unmuted checkbox
-    private JPanel audioPanel = new JPanel();
-    private JCheckBox audioMute = new JCheckBox("Mute");
+
 
     /**
      * Default constructor for the class. Sets everything up.
@@ -124,20 +106,19 @@ public class BattleshipGUI extends JFrame{
 
         //Add game board
         this.board.setSize(100,210);
-        this.board.addMouseListener(this.new cellClick());
-        this.board.addMouseMotionListener(this.new mouseMove());
-        this.board.addKeyListener(this.new changeOrientation());
+        this.board.addMouseListener(board.new cellClick());
+        this.board.addMouseMotionListener(board.new mouseMove());
+        this.board.addKeyListener(board.new changeOrientation());
         this.getContentPane().add(BorderLayout.CENTER,board);
         this.getContentPane().setBackground(Color.WHITE);
 
-        //Add audio controls
+        //Add  controls
         audioPanel.setLayout(new BorderLayout());
         audioPanel.setBackground(Color.WHITE);
         this.getContentPane().add(BorderLayout.EAST,audioPanel);
-        audioMute.setBackground(Color.WHITE);
-        audioPanel.add(audioMute, BorderLayout.SOUTH);
-        audioMute.addItemListener(new audioCheck());
-
+        board.audioMute.setBackground(Color.WHITE);
+        audioPanel.add(board.audioMute, BorderLayout.SOUTH);
+        board.audioMute.addItemListener(board.new audioCheck());
 
         //Add messages
         this.getContentPane().add(BorderLayout.SOUTH, messages);
@@ -248,12 +229,12 @@ public class BattleshipGUI extends JFrame{
         this.ipPopUp.getContentPane().add(BorderLayout.SOUTH,ipMessage);
         
         //Initialize sound file locations
-        placeURL = this.getClass().getResource("sfx/ship_place.aiff");
+        board.placeURL = this.getClass().getResource("sfx/ship_place.aiff");
         shotURL = this.getClass().getResource("sfx/ship_hit.aiff");
         missURL = this.getClass().getResource("sfx/miss_splash.aiff");
         winURL = this.getClass().getResource("sfx/victory.aiff");
         loseURL = this.getClass().getResource("sfx/failure.aiff");
-        cantPlaceURL = this.getClass().getResource("sfx/ship_cant_place.aiff");
+        board.cantPlaceURL = this.getClass().getResource("sfx/ship_cant_place.aiff");
     }
 
     /**
@@ -262,79 +243,29 @@ public class BattleshipGUI extends JFrame{
      **/
         public static void main(String[] args){
 
-            Player human = new Player();
-            human.randomGenerateBoats();
-            BattleshipGUI gui = new BattleshipGUI();
+            // Player human = new Player();
+            // human.randomGenerateBoats();
+            // BattleshipGUI gui = new BattleshipGUI();
 
-            gui.setOptions();
-            gui.setVisible(true);
+            // gui.setOptions();
+            // gui.setVisible(true);
 
-            ArrayList<Integer> boats = human.getBoatsArrayList();
-            gui.addPlayerBoats(boats);
+            // ArrayList<Integer> boats = human.getBoatsArrayList();
+            // gui.addPlayerBoats(boats);
 
-            Computer computer = new Computer("HARD");
-            ArrayList<Integer> boatList = computer.getBoatsArrayList();
-            gui.addEnemyBoats(boatList);
+            // Computer computer = new Computer("HARD");
+            // ArrayList<Integer> boatList = computer.getBoatsArrayList();
+            // gui.addEnemyBoats(boatList);
 
-            for(int i = 0; i<20; i++){
-                int move = computer.makeMove();
-                gui.addShot(move);
-                computer.updateGuessGrid( move, gui.hitEnemy(move));
-            }
+            // for(int i = 0; i<20; i++){
+            //     int move = computer.makeMove();
+            //     gui.addShot(move);
+            //     computer.updateGuessGrid( move, gui.hitEnemy(move));
+            // }
 
         }
 
-        /**
-         * Returns whether a shot from the player against the enemy is a "HIT" or a "MISS".
-         * @param shot The player's shot.
-         * @return "HIT" or "MISS"
-         **/
-        public String hitEnemy(int shot){
-            if( enemyBoats.contains(shot)) return "HIT";
-            else return "MISS";
-        }
 
-        /**
-         * Returns whether a shot from the enemy against the player is a "HIT" or a "MISS".
-         * @param shot The enemy's shot.
-         * @return "HIT" or "MISS"
-         **/
-
-        public String hitPlayer(int shot){
-            if( playerBoats.contains(shot)) return "HIT";
-            else return "MISS";
-        }
-
-        public void placeBoats() {
-            this.placeBoats = true;
-            int[] boats = Player.shipSizes; 
-            for(int boat: boats){
-                this.boatPlaced = false;
-                this.boatToPlace = boat;
-                while(!this.boatPlaced){
-                    /*try{
-                      Thread.sleep(10);
-                      }
-                      catch (InterruptedException e){
-                      }*/
-                    BattleshipController.sleep();
-                }
-            }
-            this.placeBoats = false;
-            this.repaint();
-        }
-
-        /**
-         * Shifts some location to player's GameGrid by moving it down 11 rows.
-         * @param loc The location to be shifted.
-         * @return shifted integer location value
-         **/
-
-        public int shiftToPlayerGrid(int loc){
-            int boatRow = loc/10 + 11;
-            int boatColumn = loc%10;
-            return boatRow*10 + boatColumn;
-        }
 
         /**
          * Method to initiate option prompts to the user.
@@ -353,16 +284,11 @@ public class BattleshipGUI extends JFrame{
         public void reset(){
             this.difficulty = null;
             this.gameType = 0;
-            this.playersTurn = false;
             this.ipEntered = false;
             this.replay = true;
             this.prompt = true;
-            this.lastMove = 0;
 
-            this.playerBoats = new ArrayList<Integer>();
-            this.enemyBoats = new ArrayList<Integer>();
-            this.shots= new ArrayList<Integer>();
-
+            board.reset();
             this.setVisible(false);
         }
 
@@ -370,15 +296,11 @@ public class BattleshipGUI extends JFrame{
         public void resetPlace() {
             this.difficulty = null;
             this.prompt = true;
-            this.lastMove = 0;
             this.gameType = 3;
-            this.playersTurn = false;
             this.ipEntered = false;
             this.replay = true;
              
-            this.playerBoats = new ArrayList<Integer>();
-            this.enemyBoats = new ArrayList<Integer>();
-            this.shots = new ArrayList<Integer>();
+            board.reset();
         
             this.setVisible(false);
             this.diffPopUp.setVisible(true);
@@ -388,14 +310,11 @@ public class BattleshipGUI extends JFrame{
         public void resetForIP() {
             this.difficulty = "INTERNET";
             this.prompt = true;
-            this.lastMove = 0;
             this.gameType = 2;
             this.ipEntered = false;
             this.replay = true;
              
-            this.playerBoats = new ArrayList<Integer>();
-            this.enemyBoats = new ArrayList<Integer>();
-            this.shots = new ArrayList<Integer>();
+            board.reset();
         
             this.setVisible(false);
             this.ipPopUp.setVisible(true);
@@ -406,14 +325,11 @@ public class BattleshipGUI extends JFrame{
             this.setVisible(false);
             this.difficulty = "INTERNET";
             this.prompt = true;
-            this.lastMove = 0;
             this.gameType = 2;
             this.ipEntered = false;
             this.replay = true;
              
-            this.playerBoats = new ArrayList<Integer>();
-            this.enemyBoats = new ArrayList<Integer>();
-            this.shots = new ArrayList<Integer>();
+            board.reset();
         
             this.setVisible(true);
         
@@ -423,14 +339,11 @@ public class BattleshipGUI extends JFrame{
             this.setVisible(false);
             this.difficulty = "INTERNET";
             this.prompt = true;
-            this.lastMove = 0;
             this.gameType = 1;
             this.ipEntered = false;
             this.replay = true;
              
-            this.playerBoats = new ArrayList<Integer>();
-            this.enemyBoats = new ArrayList<Integer>();
-            this.shots = new ArrayList<Integer>();
+            board.reset();
         
             this.setVisible(true);
         
@@ -439,16 +352,12 @@ public class BattleshipGUI extends JFrame{
         public void resetShips() {
             this.difficulty = null;
             this.prompt = true;
-            this.lastMove = 0;
             this.gameType = 3; 
-            this.playersTurn = false;
             this.ipEntered = false;
             this.replay = true;
 
 
-            this.playerBoats = new ArrayList<Integer>();
-            this.enemyBoats = new ArrayList<Integer>();
-            this.shots = new ArrayList<Integer>();
+            board.reset();
 
             this.setVisible(false);
             this.shipSizePopUp.setVisible(true);
@@ -483,15 +392,6 @@ public class BattleshipGUI extends JFrame{
         }
 
         /**
-         * Setter for playersTurn variable. Used by controller to let user make a move.
-         * @param tf The boolean value to set to playersTurn.
-         **/
-
-        public void setPlayersTurn(boolean tf){
-            this.playersTurn = tf;
-        }
-	
-        /**
          * Getter for replay instance variable. Used to check if the player wants to keep playing.
          * @return true for keep playing false for stop playing
          **/
@@ -516,15 +416,6 @@ public class BattleshipGUI extends JFrame{
 
         public String getDifficulty(){
             return this.difficulty;
-        }
-
-        /**
-         * Getter for playersTurn instance variable
-         * @return value stored in playersTurn
-         */
-
-        public boolean getPlayersTurn(){
-            return this.playersTurn;
         }
 
 	/**
@@ -563,209 +454,41 @@ public class BattleshipGUI extends JFrame{
 		return this.prompt;
 	 }
 	 
-	/**
-	 * Lets gui know its players turn
-	 */
-	
-	public void makeMove(){
-		this.playersTurn = true;
-	}
-	
-	/**
-	 * Check if a spawn is valid for placing boat
-	 * @param spawn representing the specific spawn
-	 */
-
-	public boolean isValidSpawn(int spawn){
-		if(this.horzOrVert){
-			for(int i=0; i < boatToPlace; i++){
-				if((spawn + i)%10 > 9 || playerBoats.contains(spawn+i) || (spawn + i)/10 != spawn/10){
-                    playAudioFile(cantPlaceURL);
-                    return false;
-                    
-                }
-			}
-		}
-		else{
-			for(int i=0; i<boatToPlace; i++){
-				if((spawn+10*i)/10 > 20 || playerBoats.contains(spawn+10*i)) {
-                    playAudioFile(cantPlaceURL);
-                    return false;
-                }
-			}
-		}
-		return true;
-	}
-	
-	public void placeBoat(int spawn){
-	    ArrayList<Integer> spawns = new ArrayList<Integer>();
-		if(horzOrVert){
-		    for(int i=0; i<this.boatToPlace; i++){
-				this.playerBoats.add(spawn + i);
-				spawns.add(spawn + i);
-		    }
-		}
-		else{
-			for(int i=0; i<this.boatToPlace; i++){
-				this.playerBoats.add(spawn + 10*i);
-				spawns.add(spawn + 10*i);
-			}
-		}
-		this.playerBoatGroups.add(spawns);
-	}
-	
-	/**
-	 * Add a shot to the gui's shots list
-	 * @param shot The shot to be added
-	 */
-	
-	public void addShot(int shot){
-		this.shots.add(shot);
-		this.repaint();
-	}
-	
-	/**
-	 * Adds locations for player's boats to the playerBoats list. Shifts their integer locations.
-	 * @param boatList A list of boat locations.
-	 */
-	
-	public void addPlayerBoats(ArrayList<Integer> boatList){
-		for( Integer loc: boatList)
-			this.playerBoats.add(shiftToPlayerGrid(loc));
-	}
-	
-	/**
-	 * Adds locations for enemy's boats to the enemyBoats list.
-	 * @param boatList A list of boat locations.
-	 */
-	
-	public void addEnemyBoats(ArrayList<Integer> boatList){
-		this.enemyBoats = boatList;
-	}
-	
-	/**
-	 * Adds a single boat location to enemyBoats
-	 **/
-	
-	public void addEnemyBoat(int boatLoc){
-		this.enemyBoats.add(boatLoc);
-	}
-	
-	/**
-	 * Method for retrieving player boats. Used when GUI is used to place boats.
-	 **/
-	
-	public ArrayList<Integer> getPlayerBoats(){
-		return this.playerBoats;
-	}
-
-    /**
-     *controller class uses this method to set a player's boat list array
-     **/
-    public ArrayList<ArrayList<Integer>> getGroupBoats(){
-	return this.playerBoatGroups;
-    }
-	
-	/**
-	 * Returns the player's most recent move.
-	 **/
-	
-	public int getLastMove(){
-		return this.lastMove;
-	}
     
     public void end() {
         this.setVisible(false);
         this.getContentPane().removeAll();
     }
 	
-	/**
-	 * Inner class that paints the literal GameGrid for the game.
-	 **/
-	
-    public class GameGrid extends JComponent{
-        public int width;
-        public int height;
-        public int cellWidth;
-        public int startX;
-
-        @Override
-            public void paintComponent(Graphics g)
-            {
-
-                width = this.getWidth() + 50; //50 added to accomodate for mute box
-                height = this.getHeight();		
-                cellWidth = height/21;
-                startX = (width - (cellWidth*10))/2;
-                Graphics2D g2d = (Graphics2D) g;
-                Color ocean = new Color(0,119,190);
-
-                //Make the background white
-                g2d.setColor(Color.WHITE);
-                g2d.fillRect(0,0,this.getWidth(),this.getHeight());
-
-                //Paint individual cells
-                for(int i=0; i < 10; i++){
-                    for(int j = 0; j<21; j++){
-                        int loc = j*10 + i;
-                        if(j==10)
-                            g2d.setColor(Color.BLACK);
-                        else if(shots.contains(loc) && (playerBoats.contains(loc) || enemyBoats.contains(loc))){
-                            g2d.setColor(Color.RED);
-                        }
-                        else if(playerBoats.contains(loc))
-                            g2d.setColor(shipColor);
-                        else if(shots.contains(loc)){
-                            g2d.setColor(ocean.darker());
-                        }
-                        else g2d.setColor(ocean);
-
-                        g2d.fillRect(startX + (i*cellWidth),j*cellWidth,cellWidth,cellWidth);
-
-                    }
-                }
-
-                //Paint GameGrid lines
-                g2d.setColor(Color.GRAY);
-                g2d.drawLine(startX,0,startX,cellWidth*21); //Far left border
-                g2d.drawLine(startX + 10*cellWidth,0,startX + 10*cellWidth,cellWidth*21); //Far right border
-                for(int i=1; i<10; i++)
-                    g2d.drawLine(startX + i*cellWidth,0,startX + cellWidth*i,cellWidth*10);	
-                for(int i=1; i<10; i++)
-                    g2d.drawLine(startX + i*cellWidth,cellWidth*11,startX + cellWidth*i,cellWidth*21);	
-                for(int j=0; j<22; j++)
-                    g2d.drawLine(startX, j*cellWidth, startX + 10*cellWidth, j*cellWidth);
-
-                //Paint boat to be placed if boats are being placed
-                int topLeftX = xLoc - cellWidth/2;
-                int topLeftY = yLoc - cellWidth/2;
-                if(placeBoats && horzOrVert){ //Draw boat to be placed horizontally
-                    g2d.setColor(shipColor);
-                    g2d.fillRect(topLeftX, topLeftY, boatToPlace*cellWidth, cellWidth);
-
-                    g2d.setColor(Color.GRAY);
-                    g2d.drawLine( topLeftX, topLeftY, topLeftX + cellWidth*boatToPlace, topLeftY);
-                    g2d.drawLine( topLeftX, topLeftY + cellWidth, topLeftX + cellWidth*boatToPlace, topLeftY + cellWidth);
-                    for(int i =0; i<boatToPlace + 1; i++){
-                        g2d.drawLine( topLeftX + i*cellWidth, topLeftY, topLeftX + i*cellWidth, topLeftY + cellWidth);
-                    }
-                }
-                else if(placeBoats){ //Draw boat to be places vertically 
-                    g2d.setColor(shipColor);
-                    g2d.fillRect( topLeftX, topLeftY, cellWidth, boatToPlace*cellWidth);
-
-                    g2d.setColor(Color.GRAY);
-                    g2d.drawLine( topLeftX, topLeftY, topLeftX, topLeftY + boatToPlace*cellWidth);
-                    g2d.drawLine( topLeftX + cellWidth, topLeftY, topLeftX + cellWidth, topLeftY + boatToPlace*cellWidth);
-                    for(int i = 0; i< boatToPlace + 1; i++){
-                        g2d.drawLine( topLeftX, topLeftY + i*cellWidth, topLeftX + cellWidth, topLeftY + i*cellWidth);
-                    }
-                }
-            }
-
-
+	public void placeBoats(){
+        board.placeBoats();
     }
-
+    public ArrayList<Integer> getPlayerBoats(){
+        return board.getPlayerBoats();
+    }
+    public void addPlayerBoats(ArrayList<Integer> boatList){
+        board.addPlayerBoats(boatList);
+    }
+    public void addEnemyBoats(ArrayList<Integer> boatList){
+        board.addEnemyBoats(boatList);
+    }
+    public void makeMove(){
+        board.makeMove();
+    }
+    public boolean getPlayersTurn(){
+        return board.getPlayersTurn();
+    }
+    public int getLastMove(){
+        return board.getLastMove();
+    }
+    public ArrayList<Integer> getEnemyBoats() {return board.getEnemyBoats();}
+    public void playAudioFile(URL audioURL){    board.playAudioFile(audioURL);}
+    public int shiftToPlayerGrid(int loc){
+        return board.shiftToPlayerGrid(loc);
+    }
+    public void addShot(int shot){ board.addShot(shot);}
+    public String hitPlayer(int shot){ return board.hitPlayer(shot);}
+    public void addEnemyBoat(int boatLoc){ board.addEnemyBoat(boatLoc);}
     /**
 	 * Listener for difficulty option buttons
 	 **/
@@ -827,6 +550,7 @@ public class BattleshipGUI extends JFrame{
 		public void actionPerformed(ActionEvent e){
 			if(e.getSource() == BattleshipGUI.this.reminder){
 				int [] inputSizes = Player.shipSizes; 
+                board.setShipSizes(inputSizes);
 				boolean isValid = true;
 				
 				for(int i=0; i<5; i++){
@@ -843,20 +567,6 @@ public class BattleshipGUI extends JFrame{
 		}
 	}
 
-    /**
-    * Listener for the mute check box
-    * audio is muted when checked and unmuted when unchecked
-    **/
-
-    public class audioCheck implements ItemListener{
-        public void itemStateChanged(ItemEvent e){
-            JCheckBox cb = (JCheckBox) e.getSource();
-            if(cb.isSelected())
-                audio = false;
-            else
-                audio = true;
-        }
-    }
 
     /**
      * Listener for the color selection menu's continue button
@@ -883,28 +593,28 @@ public class BattleshipGUI extends JFrame{
             JComboBox combo = (JComboBox)e.getSource();
             switch((String)combo.getSelectedItem()){
                 case "Default":
-                    shipColor = Color.DARK_GRAY;
+                    board.setShipColor(Color.DARK_GRAY);
                     break;
                 case "Black":
-                    shipColor = Color.BLACK;
+                    board.setShipColor(Color.BLACK);
                     break;
                 case "Blue":
-                    shipColor = Color.BLUE;
+                    board.setShipColor(Color.BLUE);
                     break;
                 case "Green":
-                    shipColor = Color.GREEN;
+                    board.setShipColor(Color.GREEN);
                     break;
                 case "Orange":
-                    shipColor = Color.ORANGE;
+                    board.setShipColor(Color.ORANGE);
                     break;
                 case "Pink":
-                    shipColor = Color.PINK;
+                    board.setShipColor(Color.PINK);
                     break;
                 case "White":
-                    shipColor = Color.WHITE;
+                    board.setShipColor(Color.WHITE);
                     break;
                 case "Yellow":
-                    shipColor = Color.YELLOW;
+                    board.setShipColor(Color.YELLOW);
                     break;
                 default:
                     break;
@@ -955,7 +665,7 @@ public class BattleshipGUI extends JFrame{
                 BattleshipGUI.this.networkPlayAgainPopUp.setVisible(false);
             }
 
-	    else if(e.getSource() == BattleshipGUI.this.playAgainAsHostButton) {
+	       else if(e.getSource() == BattleshipGUI.this.playAgainAsHostButton) {
                 replayType = 6;
                 BattleshipGUI.this.setVisible(false);
                 BattleshipGUI.this.networkPlayAgainPopUp.setVisible(false);
@@ -991,76 +701,5 @@ public class BattleshipGUI extends JFrame{
 		}
 	}
 	
-	/**
-	 * Mouse listener for clicking on game cells
-	 **/
-    public class cellClick implements MouseListener{
-		public void mouseClicked(MouseEvent e){}
-		public void mousePressed(MouseEvent e){
-			//Calculate which column & row the mouse was clicked in
-			int cellColumn = (int) Math.floor((e.getX() - board.startX)/board.cellWidth);
-			int cellRow = (int) Math.floor(e.getY()/board.cellWidth);
-			//Add to shot list if it was in enemy territory
-			if(cellRow < 10 && cellRow >=0 && cellColumn >=0 && cellColumn < 10 && playersTurn && (!shots.contains(cellRow*10 + cellColumn))){
-				lastMove = cellRow*10 + cellColumn;
-				shots.add(lastMove);
-				playersTurn = false;
-			}
-			//Record click for boat placement if it was in player's territory
-			else if( cellRow < 21 && cellRow > 10 && cellColumn >=0 && cellColumn < 10){
-				if(placeBoats){
-					int spawn = cellRow*10 + cellColumn;
-					if(isValidSpawn(spawn)){ //Place the boat on click if the spawn location is valid
-						boatPlaced = true;
-						placeBoat(spawn);
-    
-                        if(audio)
-                            playAudioFile(placeURL);
-    
-						BattleshipGUI.this.repaint();
-					}
-				}
-			}
-			repaint();
-		}
-		public void mouseReleased(MouseEvent e){}
-		public void mouseEntered(MouseEvent e){}
-		public void mouseExited(MouseEvent e){BattleshipGUI.this.repaint();}
-    }
 	
-	public class mouseMove implements MouseMotionListener{
-		public void mouseMoved(MouseEvent e){
-			if(placeBoats){ //Records x & y locations if player is placing boats
-				xLoc = e.getX();
-				yLoc = e.getY();
-				BattleshipGUI.this.repaint();
-			}
-		}
-		public void mouseDragged(MouseEvent e){}
-	}
-
-	public class changeOrientation implements KeyListener{
-		public void keyPressed(KeyEvent e){
-			horzOrVert = !horzOrVert; //Switch from drawing boats horizontally to drawing them vertically
-			BattleshipGUI.this.repaint();
-		}	
-		public void keyReleased(KeyEvent e){}
-		public void keyTyped(KeyEvent e){}
-	}
-
-    /**  
-    * Methods that plays audio clip referenced by audioURL
-    **/
-    public void playAudioFile(URL audioURL){
-        try{
-            AudioInputStream audioStream = AudioSystem.getAudioInputStream(audioURL);
-            this.clip = AudioSystem.getClip();
-            this.clip.open(audioStream);
-            clip.start();
-            clip.setMicrosecondPosition(0);
-        } catch(Exception e) {
-            System.err.println(e);
-        } 
-    }
-
 }
